+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## v1.5.0
+## v1.6.0
 Basic Functionality for 3D Shapes
+
+### v1.6.0 [8/20/18]
+#### Additions
+- Added Triangular Prism Functions Using Heron's Formula.
+
+#### Fixes
+- **BREAKING CHANGE:** `triangularPrism.volume` has been renamed to `triangularPrism.volumeByBaseAndHeight`.
+- **BREAKING CHANGE:** `triangularPrism.surfaceArea` has been renamed to `triangularPrism.surfaceAreaByBaseAndHeight`.
 
 ### v1.5.0 [8/20/18]
 #### Additions